Extract record storage into src/core/records.ts
Move verification record read/write, user pointers, global index
maintenance, audit log storage, and storage metering into a dedicated
module. Typecheck, 167 tests, and build all pass.
